api_service/requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-storage
prophet>=1.1.5
pandas
